Refactor : 객체 타입 변경 Long -> Integer

+++ src/main/java/lotto/model/InputMoney.java
@@ -12,7 +12,7 @@ public class InputMoney extends InputNumber {
     }
 
     @Override
-    public Long sendInputData() {
+    public Integer sendInputData() {
         return this.number % 1000;
     }
